# core-common-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/organization
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/branches/{branchId}/hours/{hoursId}:
    get:
      tags:
        - Branch Hours Management
      summary: Get branch hours by ID
      description: Returns operating hours of a specific branch based on its ID
      operationId: getBranchHoursById
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: string
            format: uuid
        - name: hoursId
          in: path
          description: ID of the hours to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved branch hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '404':
          description: Branch or hours not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
    put:
      tags:
        - Branch Hours Management
      summary: Update branch hours
      description: Updates existing operating hours of a specific branch with the provided details
      operationId: updateBranchHours
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: string
            format: uuid
        - name: hoursId
          in: path
          description: ID of the hours to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchHoursDTO'
        required: true
      responses:
        '200':
          description: Branch hours successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '400':
          description: Invalid branch hours data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '404':
          description: Branch or hours not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
    delete:
      tags:
        - Branch Hours Management
      summary: Delete branch hours
      description: Deletes operating hours of a specific branch based on its ID
      operationId: deleteBranchHours
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: string
            format: uuid
        - name: hoursId
          in: path
          description: ID of the hours to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Branch hours successfully deleted
        '404':
          description: Branch or hours not found
        '500':
          description: Internal server error
...
  /api/v1/branches/{branchId}/hours/filter:
    post:
      tags:
        - Branch Hours Management
      summary: Get all hours for a branch with filtering
      description: Returns a paginated list of operating hours for a specific branch based on filter criteria
      operationId: filterBranchHours
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBranchHoursDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved branch hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchHoursDTO'
        '404':
          description: Branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchHoursDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchHoursDTO'
  /api/v1/branches/{branchId}/hours:
    post:
      tags:
        - Branch Hours Management
      summary: Create new operating hours for a branch
      description: Creates new operating hours for a specific branch with the provided details
      operationId: createBranchHours
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchHoursDTO'
        required: true
      responses:
        '201':
          description: Branch hours successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '400':
          description: Invalid branch hours data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '404':
          description: Branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchHoursDTO'
...
components:
  schemas:
...
    BranchHoursDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        branchId:
          type: string
          format: uuid
        dayOfWeek:
          type: string
          enum:
            - MONDAY
            - TUESDAY
            - WEDNESDAY
            - THURSDAY
            - FRIDAY
            - SATURDAY
            - SUNDAY
        openTime:
          $ref: '#/components/schemas/LocalTime'
        closeTime:
          $ref: '#/components/schemas/LocalTime'
        isClosed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestBranchHoursDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BranchHoursDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for branch hours
    PaginationResponseBranchHoursDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BranchHoursDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...